requirements.txt
langchain
langchain-community
langchain_groq
langchain_openai
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
setuptools